% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_confidence_assessment}
\alias{get_confidence_assessment}
\title{Confidence assessment data}
\usage{
get_confidence_assessment(iso3, ci = idmc_api_key())
}
\arguments{
\item{iso3}{Character. The country ISO3 code}

\item{ci}{A character vector of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame} or NULL if the API returns nothing
}
\description{
Returns the confidence assessment data for a country by type of displacement
and fact type
}
\examples{
\dontrun{
get_confidence_assessment(iso3 = 'VNM')
get_confidence_assessment(iso3 = 'ABF')
}
}
\author{
An Chu
}
