% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_countries}
\alias{get_countries}
\title{Countries}
\usage{
get_countries(sort = c("name", "iso3"), ci = idmc_api_key())
}
\arguments{
\item{sort}{Sort the \code{data.frame} by either 'iso3' or 'name', defaults
to 'name'.}

\item{ci}{A character of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame} of search results, with two columns: 'iso3' and
    'name'.
}
\description{
Get a list of countries in the database
}
\examples{
\dontrun{
## sorting by country name (default)
get_countries()
## sorting data frame by country's ISO3 code
get_countries(sort = 'iso3')
}
}
\author{
An Chu
}
