% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_disaster_events}
\alias{get_disaster_events}
\title{Disaster Events}
\usage{
get_disaster_events(iso3, ci = idmc_api_key())
}
\arguments{
\item{iso3}{Character. The country ISO3 code}

\item{ci}{A character vector of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame} or NULL if the API returns nothing
}
\description{
Returns events associated with an iso code
}
\examples{
\dontrun{
get_disaster_events(iso3 = 'VNM')
get_disaster_events(iso3 = 'ABF')
}
}
\author{
An Chu
}
