% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_displacement_data}
\alias{get_displacement_data}
\title{Displacement data}
\usage{
get_displacement_data(iso3, year, limit, ci = idmc_api_key())
}
\arguments{
\item{iso3}{Character. The country ISO3 code}

\item{year}{Integer. The year of the records (default all years, it also
accept ranges)}

\item{limit}{Integer. Limit the number of records returned by the API}

\item{ci}{A character vector of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame}
}
\description{
Returns all available displacement data for both disaster and conflict by
year and country
}
\author{
An Chu
}
