% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_figure_analysis}
\alias{get_figure_analysis}
\title{Public figure analysis text}
\usage{
get_figure_analysis(iso3, ci = idmc_api_key())
}
\arguments{
\item{iso3}{Character. The country ISO3 code}

\item{ci}{A character vector of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame} or NULL if the API returns nothing
}
\description{
Returns the figure analysis for a country
}
\author{
An Chu
}
