% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_overview}
\alias{get_overview}
\title{Country overview}
\usage{
get_overview(iso3, ci = idmc_api_key())
}
\arguments{
\item{iso3}{The country ISO3 code}

\item{ci}{A character vector of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame}
}
\description{
Get a country displacement information and includes population and refugee
numbers
}
\examples{
\dontrun{
get_overview(iso3 = 'RUS')
get_overview(iso3 = 'VNM')
}
}
\author{
An Chu
}
