% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_strata_data}
\alias{get_strata_data}
\title{Conflict stock by latest update}
\usage{
get_strata_data(iso3, ci = idmc_api_key())
}
\arguments{
\item{iso3}{Character. The country ISO3 code}

\item{ci}{A character vector of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame}
}
\description{
Returns the strata data for a country by type of displacement and fact type
}
\author{
An Chu
}
