% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc-api-key.R
\name{idmc_api_key}
\alias{idmc_api_key}
\title{Get or set IDMC_API_KEY value}
\usage{
idmc_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new IDMC API key for the current environment?}
}
\value{
A character vector containing the IDMC API key
}
\description{
The API wrapper functions in this package all rely on a IDMC API key
residing in the environment variable `IDMC_API_KEY`. The easiest way to
accomplish this is to set it in the `.Renviron` file in your home directory.
}
\examples{
\dontrun{
idmc_api_key()
idmc_api_key(force = TRUE)
}
}
\author{
An Chu
}
