% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmcapi.R
\docType{package}
\name{idmcapi-package}
\alias{idmcapi}
\alias{idmcapi-package}
\title{idmcapi: Tools to Access Data from the Internal Displacement Monitoring Centre}
\description{
Provide API access to IDMC database which includes data about
displacement events (conflicts, disasters) through a public API.
}
\author{
\strong{Maintainer}: Chu An \email{chuvanan.cva@gmail.com}

}
\keyword{internal}
