% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_aggregated_disaster_data}
\alias{get_aggregated_disaster_data}
\title{Disaster aggregated data}
\usage{
get_aggregated_disaster_data(iso3, year, limit, ci = idmc_api_key())
}
\arguments{
\item{iso3}{Character. The country ISO3 code}

\item{year}{Integer. The year of the records (default all years, it also
accept ranges)}

\item{limit}{Integer. Limit the number of records returned by the API}

\item{ci}{A character vector of IDMC_API_KEY. See [idmc_api_key()].}
}
\value{
A \code{data.frame}
}
\description{
Returns all available disaster displacement data by country aggregating all
the events
}
\examples{
\dontrun{
## retrieve all data of Vietnam (VNM)
get_aggregated_disaster_data(iso = 'VNM')
## retrieve data from Vietnam in 2015
get_aggregated_disaster_data('VNM', year = 2015)
## retrieve data from Vietnam in period 2011-2015
get_aggregated_disaster_data('VNM', year = c(2011, 2015))
## limit the number of records returned by 10
get_aggregated_disaster_data('VNM', limit = 10)
}
}
\author{
An Chu
}
